+++ CHANGES.rst
@@ -2,7 +2,7 @@ Changelog
 =========
 
 
-1.0.9 (unreleased)
+1.0.9 (2022-05-02)
 ------------------
 
 - Remove duplicate / useless new icons & change default workinfos icon